requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
python-dotenv==1.0.0
